+++ test/pagy/calendar_test.rb
@@ -4,7 +4,9 @@ require_relative '../test_helper'
 require 'pagy/calendar'
 
 def pagy(unit: :month, **vars)
-  default = { period: [Time.new(2021, 10, 21, 13, 18, 23, 0), Time.new(2023, 11, 13, 15, 43, 40, 0)] }
+  Time.zone = 'Eastern Time (US & Canada)'
+  default = { period: [Time.zone.local(2021, 10, 21, 13, 18, 23, 0), Time.zone.local(2023, 11, 13, 15, 43, 40, 0)],
+              first_wday: :sunday }
   Pagy::Calendar.send(:create, unit, default.merge(vars))
 end
 
@@ -31,91 +33,91 @@ describe 'pagy/calendar' do
   describe 'it computes date variables for page 1 (default)' do
     it 'computes variables for :year' do
       p = pagy(unit: :year)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024)
-      _(p.from.utc).must_equal Time.gm(2021)
-      _(p.to.utc).must_equal Time.gm(2022)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024)
+      _(p.from).must_equal Time.zone.local(2021)
+      _(p.to).must_equal Time.zone.local(2022)
       _(p.pages).must_equal 3
       _(p.last).must_equal 3
     end
     it 'computes variables for :year desc' do
       p = pagy(unit: :year, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024)
-      _(p.from.utc).must_equal Time.gm(2023)
-      _(p.to.utc).must_equal Time.gm(2024)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024)
+      _(p.from).must_equal Time.zone.local(2023)
+      _(p.to).must_equal Time.zone.local(2024)
       _(p.pages).must_equal 3
       _(p.last).must_equal 3
     end
     it 'computes variables for :quarter' do
       p = pagy(unit: :quarter)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024, 1, 1)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 1)
-      _(p.to.utc).must_equal Time.gm(2022, 1, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024, 1, 1)
+      _(p.from).must_equal Time.zone.local(2021, 10, 1)
+      _(p.to).must_equal Time.zone.local(2022, 1, 1)
       _(p.pages).must_equal 9
       _(p.last).must_equal 9
     end
     it 'computes variables for :quarter' do
       p = pagy(unit: :quarter, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024, 1, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 10, 1)
-      _(p.to.utc).must_equal Time.gm(2024, 1, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024, 1, 1)
+      _(p.from).must_equal Time.zone.local(2023, 10, 1)
+      _(p.to).must_equal Time.zone.local(2024, 1, 1)
       _(p.pages).must_equal 9
       _(p.last).must_equal 9
     end
     it 'computes variables for :month' do
       p = pagy
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 12, 1)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 1)
-      _(p.to.utc).must_equal Time.gm(2021, 11, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 12, 1)
+      _(p.from).must_equal Time.zone.local(2021, 10, 1)
+      _(p.to).must_equal Time.zone.local(2021, 11, 1)
       _(p.pages).must_equal 26
       _(p.last).must_equal 26
     end
     it 'computes variables for :month desc' do
       p = pagy(unit: :month, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 12, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 1)
-      _(p.to.utc).must_equal Time.gm(2023, 12, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 12, 1)
+      _(p.from).must_equal Time.zone.local(2023, 11, 1)
+      _(p.to).must_equal Time.zone.local(2023, 12, 1)
       _(p.pages).must_equal 26
       _(p.last).must_equal 26
     end
     it 'computes variables for :week' do
       p = pagy(unit: :week)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 17)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 19)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 17)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 24)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 17)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 19)
+      _(p.from).must_equal Time.zone.local(2021, 10, 17)
+      _(p.to).must_equal Time.zone.local(2021, 10, 24)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :week with offset: 1 (Monday)' do
-      p = pagy(unit: :week, offset: 1)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 18)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 20)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 18)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 25)
+      p = pagy(unit: :week, first_wday: :monday)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 18)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 20)
+      _(p.from).must_equal Time.zone.local(2021, 10, 18)
+      _(p.to).must_equal Time.zone.local(2021, 10, 25)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :week with offset: 6 (Saturday)' do
-      p = pagy(unit: :week, offset: 6)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 16)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 18)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 16)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 23)
+      p = pagy(unit: :week, first_wday: :saturday)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 16)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 18)
+      _(p.from).must_equal Time.zone.local(2021, 10, 16)
+      _(p.to).must_equal Time.zone.local(2021, 10, 23)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :day' do
       p = pagy(unit: :day)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 21)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 14)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 21)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 22)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 21)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 14)
+      _(p.from).must_equal Time.zone.local(2021, 10, 21)
+      _(p.to).must_equal Time.zone.local(2021, 10, 22)
       _(p.pages).must_equal 754
       _(p.last).must_equal 754
     end
@@ -124,91 +126,91 @@ describe 'pagy/calendar' do
   describe 'it computes date variables for page 2' do
     it 'computes variables for :year' do
       p = pagy(unit: :year, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024)
-      _(p.from.utc).must_equal Time.gm(2022)
-      _(p.to.utc).must_equal Time.gm(2023)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024)
+      _(p.from).must_equal Time.zone.local(2022)
+      _(p.to).must_equal Time.zone.local(2023)
       _(p.pages).must_equal 3
       _(p.last).must_equal 3
     end
     it 'computes variables for :year desc' do
       p = pagy(unit: :year, page: 2, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024)
-      _(p.from.utc).must_equal Time.gm(2022)
-      _(p.to.utc).must_equal Time.gm(2023)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024)
+      _(p.from).must_equal Time.zone.local(2022)
+      _(p.to).must_equal Time.zone.local(2023)
       _(p.pages).must_equal 3
       _(p.last).must_equal 3
     end
     it 'computes variables for :quarter' do
       p = pagy(unit: :quarter, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024, 1, 1)
-      _(p.from.utc).must_equal Time.gm(2022, 1, 1)
-      _(p.to.utc).must_equal Time.gm(2022, 4, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024, 1, 1)
+      _(p.from).must_equal Time.zone.local(2022, 1, 1)
+      _(p.to).must_equal Time.zone.local(2022, 4, 1)
       _(p.pages).must_equal 9
       _(p.last).must_equal 9
     end
     it 'computes variables for :quarter' do
       p = pagy(unit: :quarter, page: 2, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024, 1, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 7, 1)
-      _(p.to.utc).must_equal Time.gm(2023, 10, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024, 1, 1)
+      _(p.from).must_equal Time.zone.local(2023, 7, 1)
+      _(p.to).must_equal Time.zone.local(2023, 10, 1)
       _(p.pages).must_equal 9
       _(p.last).must_equal 9
     end
     it 'computes variables for :month' do
       p = pagy(page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 12, 1)
-      _(p.from.utc).must_equal Time.gm(2021, 11, 1)
-      _(p.to.utc).must_equal Time.gm(2021, 12, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 12, 1)
+      _(p.from).must_equal Time.zone.local(2021, 11, 1)
+      _(p.to).must_equal Time.zone.local(2021, 12, 1)
       _(p.pages).must_equal 26
       _(p.last).must_equal 26
     end
     it 'computes variables for :month desc' do
       p = pagy(page: 2, order: :desc)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 12, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 10, 1)
-      _(p.to.utc).must_equal Time.gm(2023, 11, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 12, 1)
+      _(p.from).must_equal Time.zone.local(2023, 10, 1)
+      _(p.to).must_equal Time.zone.local(2023, 11, 1)
       _(p.pages).must_equal 26
       _(p.last).must_equal 26
     end
     it 'computes variables for :week' do
       p = pagy(unit: :week, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 17)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 19)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 24)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 31)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 17)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 19)
+      _(p.from).must_equal Time.zone.local(2021, 10, 24)
+      _(p.to).must_equal Time.zone.local(2021, 10, 31)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :week with offset: 1 (Monday)' do
-      p = pagy(unit: :week, offset: 1, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 18)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 20)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 25)
-      _(p.to.utc).must_equal Time.gm(2021, 11, 1)
+      p = pagy(unit: :week, first_wday: :monday, page: 2)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 18)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 20)
+      _(p.from).must_equal Time.zone.local(2021, 10, 25)
+      _(p.to).must_equal Time.zone.local(2021, 11, 1)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :week with offset: 6 (Saturday)' do
-      p = pagy(unit: :week, offset: 6, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 16)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 18)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 23)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 30)
+      p = pagy(unit: :week, first_wday: :saturday, page: 2)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 16)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 18)
+      _(p.from).must_equal Time.zone.local(2021, 10, 23)
+      _(p.to).must_equal Time.zone.local(2021, 10, 30)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
     end
     it 'computes variables for :day' do
       p = pagy(unit: :day, page: 2)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 21)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 14)
-      _(p.from.utc).must_equal Time.gm(2021, 10, 22)
-      _(p.to.utc).must_equal Time.gm(2021, 10, 23)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 21)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 14)
+      _(p.from).must_equal Time.zone.local(2021, 10, 22)
+      _(p.to).must_equal Time.zone.local(2021, 10, 23)
       _(p.pages).must_equal 754
       _(p.last).must_equal 754
     end
@@ -217,10 +219,10 @@ describe 'pagy/calendar' do
   describe 'it computes date variables for last page and overflow' do
     it 'computes variables for :year' do
       p = pagy(unit: :year, page: 3)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024)
-      _(p.from.utc).must_equal Time.gm(2023)
-      _(p.to.utc).must_equal Time.gm(2024)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 1, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024)
+      _(p.from).must_equal Time.zone.local(2023)
+      _(p.to).must_equal Time.zone.local(2024)
       _(p.pages).must_equal 3
       _(p.last).must_equal 3
       _(pagy(unit: :year, page: 3, cycle: true).next).must_equal 1
@@ -228,19 +230,19 @@ describe 'pagy/calendar' do
     end
     it 'computes variables for :quarter' do
       p = pagy(unit: :quarter, page: 9)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2024, 1, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 10, 1)
-      _(p.to.utc).must_equal Time.gm(2024, 1, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2024, 1, 1)
+      _(p.from).must_equal Time.zone.local(2023, 10, 1)
+      _(p.to).must_equal Time.zone.local(2024, 1, 1)
       _(p.pages).must_equal 9
       _(p.last).must_equal 9
     end
     it 'computes variables for :month' do
       p = pagy(page: 26)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 1)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 12, 1)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 1)
-      _(p.to.utc).must_equal Time.gm(2023, 12, 1)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 1)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 12, 1)
+      _(p.from).must_equal Time.zone.local(2023, 11, 1)
+      _(p.to).must_equal Time.zone.local(2023, 12, 1)
       _(p.pages).must_equal 26
       _(p.last).must_equal 26
       _(pagy(page: 26, cycle: true).next).must_equal 1
@@ -248,32 +250,32 @@ describe 'pagy/calendar' do
     end
     it 'computes variables for :week' do
       p = pagy(unit: :week, page: 109)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 17)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 19)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 12)
-      _(p.to.utc).must_equal Time.gm(2023, 11, 19)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 17)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 19)
+      _(p.from).must_equal Time.zone.local(2023, 11, 12)
+      _(p.to).must_equal Time.zone.local(2023, 11, 19)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
       _(pagy(unit: :week, page: 109, cycle: true).next).must_equal 1
       _ { pagy(unit: :week, page: 110) }.must_raise Pagy::OverflowError
     end
     it 'computes variables for :week with offset: 1 (Monday)' do
-      p = pagy(unit: :week, offset: 1, page: 109)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 18)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 20)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 13)
-      _(p.to.utc).must_equal Time.gm(2023, 11, 20)
+      p = pagy(unit: :week, first_wday: :monday, page: 109)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 18)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 20)
+      _(p.from).must_equal Time.zone.local(2023, 11, 13)
+      _(p.to).must_equal Time.zone.local(2023, 11, 20)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
       _(pagy(unit: :week, page: 109, cycle: true).next).must_equal 1
       _ { pagy(unit: :week, page: 110) }.must_raise Pagy::OverflowError
     end
     it 'computes variables for :week with offset: 6 (Saturday)' do
-      p = pagy(unit: :week, offset: 6, page: 109)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 16)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 18)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 11)
-      _(p.to.utc).must_equal Time.gm(2023, 11, 18)
+      p = pagy(unit: :week, first_wday: :saturday, page: 109)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 16)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 18)
+      _(p.from).must_equal Time.zone.local(2023, 11, 11)
+      _(p.to).must_equal Time.zone.local(2023, 11, 18)
       _(p.pages).must_equal 109
       _(p.last).must_equal 109
       _(pagy(unit: :week, page: 109, cycle: true).next).must_equal 1
@@ -281,10 +283,10 @@ describe 'pagy/calendar' do
     end
     it 'computes variables for :day' do
       p = pagy(unit: :day, page: 754)
-      _(p.instance_variable_get(:@initial)).must_equal Time.gm(2021, 10, 21)
-      _(p.instance_variable_get(:@final)).must_equal Time.gm(2023, 11, 14)
-      _(p.from.utc).must_equal Time.gm(2023, 11, 13)
-      _(p.to.utc).must_equal Time.gm(2023, 11, 14)
+      _(p.instance_variable_get(:@initial)).must_equal Time.zone.local(2021, 10, 21)
+      _(p.instance_variable_get(:@final)).must_equal Time.zone.local(2023, 11, 14)
+      _(p.from).must_equal Time.zone.local(2023, 11, 13)
+      _(p.to).must_equal Time.zone.local(2023, 11, 14)
       _(p.pages).must_equal 754
       _(p.last).must_equal 754
       _(pagy(unit: :day, page: 754, cycle: true).next).must_equal 1