+++ CHANGELOG.md
@@ -14,7 +14,8 @@ If you upgrade from version `< 5.0.0` see the following:
 
 Still supported in version `5.x` but not supported from `6.0` on:
 
-- `pagy_massage_params` method: use the `:params` variable set to a lambda `Proc` that does the same (but per instance). See [How to customize the params](https://ddnexus.github.io/pagy/how-to#customize-the-params).
+- The `pagy_massage_params` method: use the `:params` variable set to a lambda `Proc` that does the same (but per instance). See [How to customize the params](https://ddnexus.github.io/pagy/how-to#customize-the-params).
+- The `:offset` variable used by the `Pagy::Calendar::Week`: set the `:first_wday` variable to the symbol of the first day of the week (e.g.: `first_wday: :sunday`).
 
 <hr>